requirements.txt
langchain
openai 
faiss-cpu 
tiktoken 
pydantic 
fastapi 
uvicorn 
python-dotenv
langchain-community 
langchain-core
fastapi[standard]
